voice-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voice-agent"
version = "0.1.0"
description = "Ultra-low-latency interruptible voice agent using LiveKit + Pipecat + Voxtral STT + Chatterbox TTS + Cerebras LLM"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "pipecat-ai[livekit,silero]>=0.0.50",
    "openai>=1.12.0",
    "pydantic-settings>=2.1.0",
    "pydantic>=2.5.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "livekit-api>=0.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/voice_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
